Refactor 'Drinks' component
