+++ src/components/Drinks/index.tsx
@@ -1,0 +1,1 @@
+export { default as Drinks } from './Drinks';
